pyproject.toml
[project]
name = "tcc-enem-main"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.10.3",
    "numpy>=2.2.6",
    "pandas>=2.2.3",
    "scikit-learn>=1.6.1",
    "streamlit-local-storage>=0.0.25",
    "toolz>=1.0.0",
]
